backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
python-jose[cryptography]
boto3
pillow
numpy
python-multipart
websockets
aiosqlite
pydantic[email]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
starlette
opencv-python-headless